\name{patternsExample}
\alias{patternsExample}
\docType{data}
\title{patternsExample}
\description{
   A data frame which contains a column of methylation patterns and two columns
   of counts
}
\usage{data(patternsExample)}
\format{
  This data frame contains the following columns:
    \describe{
      \item{mPattern}{methylation patterns}
      \item{k1}{first column of counts}
      \item{k2}{second column of counts}
    }
}

% vim:ts=2:sw=2:sts=2:expandtab:
