\name{patternsExample}
\alias{patternsExample}
\docType{data}
\title{patternsExample}
\description{
   A data frame which contains a column of methylation patterns and two columns
   of counts. This data was obtained as described in 
   Lyko, F., Forest, S., Kucharski, R., Wolf, S., Falckenhayn, C., 
   and Maleszka, R. (2010). The honey bee epigenomes: differential methylation of brain DNA in 
   queens and workers. PLoS Biol, 8(11), e1000506.
}
\usage{data(patternsExample)}
\format{
  This data frame contains the following columns:
    \describe{
      \item{mPattern}{methylation patterns}
      \item{k1}{first column of counts}
      \item{k2}{second column of counts}
    }
}

% vim:ts=2:sw=2:sts=2:expandtab:
